编译告警整改

+++ usb/gadget/function/mtp/include/usbfn_mtp_impl.h
@@ -258,7 +258,7 @@ private:
     static sem_t asyncReq_;
     static pthread_rwlock_t mtpRunrwLock_;
     std::vector<uint8_t> vectorSplited_;
-    int32_t writeActualLen_;
+    size_t writeActualLen_;
 };
 } // namespace V1_0
 } // namespace Mtp
